Tests/IsShortArrayOrList: move original tests
This commit moves the original test set for the `IsShortArrayOrList` functionality from the `Utils\Lists` test directory to the `Internal\IsShortArrayOrList` test directory.

Includes updating the `README-IsShortArrayTests.md` file and adding a `README-IsShortListTests.md` file documenting where the tests are located.
