+++ Tests/Utils/Arrays/README-IsShortArrayTests.md
@@ -1,4 +1,4 @@
 # Regarding tests for `Arrays::isShortArray()`
 
 The `Arrays::isShortArray()` method is tested together with the `Lists::isShortList()` method.
-The tests for this can be found in the `Tests/Utils/Lists` folder.
+The tests for this can be found in the `Tests/Internal/IsShortArrayOrList` folder.
