+++ Tests/Utils/Lists/README-IsShortListTests.md
@@ -1,0 +1,4 @@
+# Regarding tests for `Lists::isShortList()`
+
+The `Lists::isShortList()` method is tested together with the `Arrays::isShortArray()` method.
+The tests for this can be found in the `Tests/Internal/IsShortArrayOrList` folder.
